chore: split `AssetRow` into multiple files
